Add files via upload
In this project, I have worked with the telecom company’s customer dataset. This dataset includes
subscribing to telephone customer’s details. Each of the columns has data on phone number, call
minutes during various times of the day, the charges incurred, lifetime account duration, and whether
the customer has churned some services by unsubscribing it. The goal is to predict whether a
customer will eventually churn or not.